TVB-2663 Move pse.py from tvb.core to tvb.adapters
